move to utils/io.py
